Initial implementation of the MessagingService
This version provides the basic messaging patterns required for
communication in a network of elevator controllers.

A separate asio execution context is currently wrapped inside the
implementation class, so that the service can be used by an application
that is not necessarily based on ASIO. In the case where it is more
widely used, it may be suitable for this service to take in an existing
io_context. Beware that this will impact the existing thread-safety.

The service can be further extended to expose more information that is
available from the zyre-network, including more detailed events and
information about available peers and topics.